getConfig runs "cabal configure" first if setup-config is invalid.
